requirements.txt
dawdreamer==0.8.0
numpy==1.26.2
scipy==1.11.4
playsound2==1.3.0
pyobjc==10.1